Move pkg/dashboard -> pkg/web
