+++ pkg/web/assets.go
@@ -1,4 +1,4 @@
-package dashboard
+package web
 
 import (
 	"net/http"
