+++ pkg/web/assets/css/check-details.css
