+++ pkg/web/assets/css/dashboard.css
