+++ pkg/web/assets/images/favicon-32x32.png
