+++ pkg/web/assets/images/favicon.ico
