+++ pkg/web/assets/webfonts/fa-regular-400.woff2
