+++ pkg/web/router.go
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package dashboard
+package web
 
 import (
 	"net/http"
