+++ pkg/web/templates/dashboard.gohtml
