+++ pkg/web/templates/navbar.gohtml
